extract logic
