+++ app.js
@@ -1,9 +1,8 @@
 #!/usr/bin/env node
 
-const _ = require('lodash');
-const moment = require('moment');
 const fs = require('fs');
 const program = require('commander');
+const { getKeys, addTime } = require('./src/premiertime');
 
 program
     .version('0.1.2')
@@ -11,30 +10,12 @@ program
 
 program.parse(process.argv);
 
+// init file names
 const inputFileName = program.args[0];
 const matcher = inputFileName.match(/([^\/]*).csv$/);
 const outputFileName = `${matcher[1]}.output.csv`;
 
-function getKeys(data) {
-    return _.uniqBy(data, 'name').map(entry => entry.name);
-}
-
-function addTime(a, b) {
-    const durA = moment.duration(a);
-    const durB = moment.duration(b);
-    const sum = durA.add(durB);
-    const hour = Math.floor(sum.asHours());
-    const minute = ('00' + sum.minutes()).slice(-2);
-
-    return `${hour}:${minute}`;
-}
-
-module.exports = {
-    getKeys: getKeys,
-    addTime: addTime
-};
-
-
+// read files
 const raw = fs.readFileSync(inputFileName).toString();
 const data = raw.split(/\s+/)
     .filter(entry => /.*,\d+:\d+/.test(entry))
@@ -64,6 +45,7 @@ const totalTime = times.reduce((acc, entry) => addTime(acc, entry.time), '0:00')
 
 console.log(times, totalTime);
 
+// write result
 const outputFd = fs.open(`./${outputFileName}`, 'w', (err, fd) => {
     times.forEach(entry => {
         fs.write(fd, `${entry.name},${entry.time}\n`, () => {});
